v0.5
Defined Mark and Maven as a network of bots, jobs, companies, and people (each of them “nodes”). 
Defined Nodes as having operating systems and "profiles."
Defined Profiles as sets of "attributes."
Defined Attributes as characteristics/traits (as-yet undefined), experiences (as-yet undefined), and relationships (as-yet undefined) to other nodes.
Defined mavenOS as the operating system for a bot in the network called Maven.